Move stringify email to utils
