+++ apps/web/utils/ai/assistant/process-user-request.ts
@@ -27,7 +27,7 @@ import {
 import { updateCategoryForSender } from "@/utils/categorize/senders/categorize";
 import { findSenderByEmail } from "@/utils/sender";
 import { getEmailForLLM } from "@/utils/ai/choose-rule/get-email-from-message";
-import { stringifyEmailSimple } from "@/utils/ai/choose-rule/stringify-email";
+import { stringifyEmailSimple } from "@/utils/stringify-email";
 import {
   updatePromptFileOnRuleCreated,
   updatePromptFileOnRuleUpdated,
